+++ HomeAccessPlusDocumentProvider/DocumentPickerViewController.swift
@@ -75,6 +75,7 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
     ///   3. The type of item this is (Drive, Directory, File Type)
     ///   4. The extension of the file, or empty if it is a directory
     ///   5. Additional details for the file (size, modified date, etc...)
+    ///   6. User has write permission for the item
     var fileItems: [NSArray] = []
     
     /// A reference to the original URL path to the file that
@@ -237,7 +238,7 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
     ///
     /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
     /// - since: 0.3.0-beta
-    /// - version: 4
+    /// - version: 5
     /// - date: 2016-02-25
     func loadFileBrowser() {
         // Hiding the built in table refresh control, as the
@@ -289,9 +290,19 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
                             // See: http://stackoverflow.com/a/25340084
                             space = String(format:"%.2f", subJson["Space"].double!) + "% used"
                         }
+                        // Seeing if the user has write permission to the drive.
+                        // This is set in the hapconfig.xml file, and through
+                        // guesswork, it seems as though the JSON value returned
+                        // from the API is called "Actions". It is "0" if write
+                        // is allowed, or "1" if it is not (I think)
+                        var writePermission = false
+                        if (subJson["Actions"].int == 0) {
+                            writePermission = true
+                        }
                         logger.debug("Drive name: \(name!)")
                         logger.debug("Drive path: \(path!)")
                         logger.debug("Drive usage: \(space)")
+                        logger.debug("Drive write permission: \(writePermission)")
                         
                         // Creating a drive letter to show under the name of the
                         // drive, as this will be familiar to how drives are
@@ -300,7 +311,7 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
                         
                         // Adding the current files and folders in the directory
                         // to the fileItems array
-                        self.addFileItem(name!, path: path!, type: driveLetter + " Drive", fileExtension: "Drive", details: space)
+                        self.addFileItem(name!, path: path!, type: driveLetter + " Drive", fileExtension: "Drive", details: space, writePermission: writePermission)
                     }
                     
                     // Hiding the HUD and adding the drives available to the table
@@ -364,13 +375,15 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
                         let size = subJson["Size"].string
                         let path = subJson["Path"].string
                         let details = modified! + "    " + size!
+                        let writePermission = subJson["Permissions"]["Write"].boolValue
                         logger.verbose("Name: \(name)")
                         logger.verbose("Type: \(type)")
                         logger.verbose("Date modified: \(modified)")
+                        logger.verbose("Write permission: \(writePermission)")
                         
                         // Adding the current files and folders in the directory
                         // to the fileItems array
-                        self.addFileItem(name!, path: path!, type: type!, fileExtension: fileExtension!, details: details)
+                        self.addFileItem(name!, path: path!, type: type!, fileExtension: fileExtension!, details: details, writePermission: writePermission)
                     }
                     self.hudHide()
                     
@@ -407,7 +420,7 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
     ///
     /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
     /// - since: 0.3.0-beta
-    /// - version: 1
+    /// - version: 2
     /// - date: 2015-12-18
     ///
     /// - seealso: fileItems
@@ -417,13 +430,14 @@ class DocumentPickerViewController: UIDocumentPickerExtensionViewController, UIT
     /// - parameter type: The type of item this is (Drive, Directory, File Type)
     /// - parameter fileExtension: The extension of the file, or empty if it is a directory
     /// - parameter details: Additional info for the file (size, modified date, etc...)
-    func addFileItem(_ name: String, path: String, type: String, fileExtension: String, details: String) {
+    // - parameter writePermission: User has write permission for the item
+    func addFileItem(_ name: String, path: String, type: String, fileExtension: String, details: String, writePermission: Bool) {
         // Creating an array to hold the current item that is being processed
         var currentItem: [AnyObject] = []
-        logger.verbose("Adding item to file array, with details:\n --Name: \(name)\n --Path: \(path)\n --Type: \(type)\n --Extension: \(fileExtension)\n --Details: \(details)")
+        logger.verbose("Adding item to file array, with details:\n --Name: \(name)\n --Path: \(path)\n --Type: \(type)\n --Extension: \(fileExtension)\n --Details: \(details)\n --Write permission: \(writePermission)")
         
         // Adding the current item to the array
-        currentItem = [name as AnyObject, path as AnyObject, type as AnyObject, fileExtension as AnyObject, details as AnyObject]
+        currentItem = [name as AnyObject, path as AnyObject, type as AnyObject, fileExtension as AnyObject, details as AnyObject, writePermission as AnyObject]
         self.fileItems.append(currentItem as NSArray)
     }