+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0
 
+### New Stuff
+* "Upload" and "Create" options in the upload popover are disabled if the user does not have write access granted for the current directory
+
 ### Bug Fixes
 * Fixed bug where the app would crash if a user tried to open the upload popover when they had logged out of the app previously, while app restoration made it appear as though they were still logged in